@@ -26,10 +26,10 @@
     "jwks-rsa": "^2.0.5",
     "mongoose": "^6.1.4",
     "postcss": "^8.4.5",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-is": "17.0.2",
-    "react-router-dom": "5.3.0",
+    "react-router-dom": "^6.3.0",
     "reflect-metadata": "^0.1.13",
     "regenerator-runtime": "0.13.7",
     "rxjs": "^7.0.0",
@@ -62,8 +62,8 @@
     "@testing-library/react-hooks": "7.0.2",
     "@types/jest": "27.0.2",
     "@types/node": "14.14.33",
-    "@types/react": "17.0.30",
-    "@types/react-dom": "17.0.9",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
     "@types/react-is": "17.0.2",
     "@types/react-router-dom": "5.3.1",
     "@typescript-eslint/eslint-plugin": "~5.3.0",
